Refactor the app into small, more focused classes
Previously, the app was built to work, but with very little thought to
optimisation and maintenance. Sure, it would have been sufficient for a
tutorial but, wasn't the best example of writing code.

This change seeks to correct that by refactoring the existing code from
public/index.php into a series of more task-specific classes, and using
the application's DI container as much and as well as possible, within
the available time constraints.