Replace Event naming with Meeting
All of the relevant references to "event" have been changed to "meeting"
to better reflect how the Zoho CRM API is structured.

While it's not much of a muchness, I feel that changing references,
where practical and relevant, to "event" to "meeting", perhaps, better
reflects the Zoho CRM API, so should be more intuitive when reading the
code.