Change the default route's allowed method
The default route's allowed method has been changed from GET to POST
which is what is required by Zoho CRM's webhooks.

I think that I set it as GET during development, as that was the default
that I'm used to working with (or, I copied some initial code to get up
and going quicker and the method was set to GET).