Migrate several variables to environment variables
During the technical review of the accompanying tutorial to this
application, the reviewer identified that some of the variables were
specific to my account or other users in the region I'm in.
Consequently, the application would not work for users outside of this
region. So, this change refactored the code to be to be more flexible.