Add a method to test call logging
This change adds the method testCallLogging as a simplistic way of
testing whether the ability to log a call with Zoho CRM works or not. I
added it as I was having issues with logging calls and wanted a way to
just focus on the call to Zoho CRM's API, without the surrounding code
for making and handling a call to Twilio.