Correct a bug logging calls
After quite some research and experimentation, I learned that the
error I was seeing was because I was marking calls as inbound, yet
providing outbound call details as well. This doesn't really explain the
mismatched error messages, but at least I understand what was going
wrong.

So, this change refactors the relevant code in src/Application.php to
log inbound calls (as that's the use case the app is centered around),
and deprecates any irrelevant code.